Moved test function to test_solve where it fits better.
