+++ respy/tests/test_simulate.py
@@ -5,42 +5,7 @@ import pytest
 
 import respy as rp
 from respy.likelihood import get_crit_func
-from respy.pre_processing.model_processing import process_params_and_options
-from respy.simulate import calculate_value_functions_and_flow_utilities
 from respy.tests.random_model import generate_random_model
-
-
-@pytest.mark.parametrize("seed", range(10))
-def test_equality_of_total_values_and_rewexpost_for_myopic_individuals(seed):
-    """Test equality of ex-post rewards and total values for myopic individuals."""
-    np.random.seed(seed)
-
-    # We need to simulate the model to get the emaxs and model attributes.
-    params, options = generate_random_model(myopic=True)
-    params, optim_paras, options = process_params_and_options(params, options)
-
-    draws = np.random.randn(1, 4)
-
-    state_space = rp.solve(params, options)
-
-    for period in range(options["n_periods"]):
-        # Unpack necessary attributes
-        wages = state_space.get_attribute_from_period("wages", period)
-        nonpec = state_space.get_attribute_from_period("nonpec", period)
-        emaxs_period = state_space.get_attribute_from_period(
-            "continuation_values", period
-        )
-        is_inadmissible = state_space.get_attribute_from_period(
-            "is_inadmissible", period
-        )
-
-        value_functions, flow_utilities = calculate_value_functions_and_flow_utilities(
-            wages, nonpec, emaxs_period, draws, optim_paras["delta"], is_inadmissible
-        )
-
-        np.testing.assert_equal(
-            value_functions[~is_inadmissible], flow_utilities[~is_inadmissible]
-        )
 
 
 @pytest.mark.parametrize("seed", range(20))
