+++ respy/tests/test_solve.py
@@ -2,18 +2,21 @@ import numpy as np
 import pytest
 from numba import njit
 
+import respy as rp
 from respy._numba import array_to_tuple
 from respy.config import EXAMPLE_MODELS
 from respy.config import KEANE_WOLPIN_1994_MODELS
 from respy.config import KEANE_WOLPIN_1997_MODELS
 from respy.pre_processing.model_checking import check_model_solution
 from respy.pre_processing.model_processing import process_params_and_options
+from respy.simulate import calculate_value_functions_and_flow_utilities
 from respy.solve import get_continuation_values
 from respy.solve import solve
 from respy.solve import StateSpace
 from respy.state_space import _create_state_space
 from respy.tests._former_code import _create_state_space_kw94
 from respy.tests._former_code import _create_state_space_kw97
+from respy.tests.random_model import generate_random_model
 from respy.tests.utils import process_model_or_seed
 
 
@@ -220,3 +223,35 @@ def test_create_state_space_vs_specialized_kw97(model_or_seed):
     mask_old = indexer_old != -1
     mask_new = indexer_new != -1
     assert np.array_equal(mask_old, mask_new)
+
+
+@pytest.mark.parametrize("seed", range(10))
+def test_equality_of_total_values_and_rewexpost_for_myopic_individuals(seed):
+    """Test equality of ex-post rewards and total values for myopic individuals."""
+    np.random.seed(seed)
+
+    # We need to simulate the model to get the emaxs and model attributes.
+    params, options = generate_random_model(myopic=True)
+    params, optim_paras, options = process_params_and_options(params, options)
+
+    draws = np.random.randn(1, 4)
+
+    state_space = rp.solve(params, options)
+
+    for period in range(options["n_periods"]):
+        wages = state_space.get_attribute_from_period("wages", period)
+        nonpec = state_space.get_attribute_from_period("nonpec", period)
+        emaxs_period = state_space.get_attribute_from_period(
+            "continuation_values", period
+        )
+        is_inadmissible = state_space.get_attribute_from_period(
+            "is_inadmissible", period
+        )
+
+        value_functions, flow_utilities = calculate_value_functions_and_flow_utilities(
+            wages, nonpec, emaxs_period, draws, optim_paras["delta"], is_inadmissible
+        )
+
+        np.testing.assert_equal(
+            value_functions[~is_inadmissible], flow_utilities[~is_inadmissible]
+        )
